02_IntroductionGolang

+++ 02_IntroductionGolang/praktikum/1Luas_trapesium.go
@@ -7,7 +7,6 @@ func main() {
 	T := 5
 	A := 12
 	B := 8
-	L := 0.5 * float32(A+B) * float32(T)
+	L := 0.5 * float64(A+B) * float64(T)
 	fmt.Println(L)
-
 }
